Delegate to the Kotlin standard library as much as possible
There will be hundreds of functions for each array type.  Implementing each of these and properly validating them will become unmaintainable so I made the decision to delegate as much as possible to the standard library.  Since many of the functions are inline, this forced me to make the backing "values" array internal with the PublishedApi annotation in order to be able to reference that from inline functions and to delegate to inline extension functions.