Fix IntelliJ IDE
After the recent IntelliJ upgrade, IntelliJ would sporadically complain about the JDK not being configured.  I think this was because IntelliJ was maintaining it's own list of modules in addition to the ones defined in gradle.

Deleting the .idea folder and re-opened the project to have the necessary files re-generated (and restored the .idea/.gitinore) and this seems to have fixed it.