Update Sequence.toImmutableArray to use builders
This is much more efficient when dealing with one of the 8 base types as it stores the primitive values in a primitive array instead of hanging onto the wrapper objects so that the memory consumption can remain low.

This will also take advantage of future builder optimizations that will improve performance for all types including generic types (I'm planning to tackle this optimization in a future release).