Rename copyOf to copyFrom
The original copyOf name suggested that it would produce an array since it would be a copy of the provided array but this produces an Immutable Array instead.  Also, The regular array copyOf function allows creating a new array that's larger than the source array by padding with the default value (null / 0) but this doesn't have that same behavior so it's best to use copyFrom instead to reduce the possibility of confusion.