Improve benchmark normalization
The previous benchmark results had the values normalized for each data type individually.  This accurately portrayed the relative performance across the 3 data structures (list, array, & immutable array) for a given data type.

However, the previous approach made it easy to jump to the wrong conclusion when considering the same data structure but operating on different data types since they were presented in the same chart but normalized separately. For example, if the performance of an Immutable-Array operation drops by a smaller amount compared to lists going from operating on Int values to Double values, the operation could be 2X faster than lists when operating on Int values and 3X faster than lists when operating on Double values.  So looking at both scenarios together in the same chart gave the false impression that working with Double values is faster than working with Int values.

This change removes this confusion by having all results within each chart normalized against the same value so that the sizes of all bars are directly comparable against all other bars within the same chart.  The new charts maintain the same relative performance across data structures but also allows us to accurately gauge the performance impact of operating on different data types.

In order to provide the most accurate feel when looking at the charts, the results are normalized against the median list performance.  In the scenario where operating on Double values is twice as slow as when operating on Int values, if the median list performance occurs when operating on Int values, then the List<Int> scenario would have a relative throughput of 1.0 and the List<Double> scenario would have a relative throughput of 0.5 so that the size of all bars are directly comparable.