Capture latest release version in json file
Shields unfortunately adds a "v" prefix to the version number and this misleads users when attempting to define the dependency version. Since that's not configurable, we need to work around that by capturing the version as json and use the dynamic json shield to fetch and extract the version from this json instead. This also has the benefit of allowing us to control when the new version is referenced in the documentation after performing some post-release sanity tests.