Create Selection class
This will be used to optimize operations that need to conditionally select elements.

The properties, such as the size and number of elements that are selected, are stored in the same backing array as the bitmap.  This allows us to use an inline class and eliminate the object overhead of this abstraction.