Optimize partition by replacing the buffer with a bitmap
The position of the 1-bits represents the passing elements and the other elements are represented by the 0-bits.

This reduces temporary memory consumption by the element size.  Eg. 32-bit references now use 32X less temporary memory as each element is represented by a single bit.

After populating the bitmap, we create perfectly-sized results, and traverse the bitmap twice in sub-linear time, first for the 1-bits and then for the 0-bits.